Moved XmlUtilities methods to XMLUtils
