+++ src/main/java/org/opengis/cite/cdb10/MetadataAndVersioning/DefaultsXmlStructureTests.java
@@ -1,8 +1,7 @@
 package org.opengis.cite.cdb10.MetadataAndVersioning;
 
 import org.opengis.cite.cdb10.CommonFixture;
-import org.opengis.cite.cdb10.util.SchemaValidatorErrorHandler;
-import org.opengis.cite.cdb10.util.XmlUtilities;
+import org.opengis.cite.cdb10.util.*;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import org.w3c.dom.Node;
@@ -30,7 +29,7 @@ public class DefaultsXmlStructureTests extends CommonFixture {
         File xmlFile = Paths.get(path, "Metadata", "Defaults.xml").toFile();
         File xsdFile = Paths.get(path, "Metadata", "Schema", "Defaults.xsd").toFile();
 
-        SchemaValidatorErrorHandler errorHandler = XmlUtilities.validateXmlFileIsValid(xmlFile, xsdFile);
+        SchemaValidatorErrorHandler errorHandler = XMLUtils.validateXmlFileIsValid(xmlFile, xsdFile);
 
         if (!errorHandler.noErrors()) {
             Assert.fail(xmlFile.getName() + " does not contain valid XML. Errors: " + errorHandler.getMessages());
@@ -39,7 +38,7 @@ public class DefaultsXmlStructureTests extends CommonFixture {
 
     @Test
     public void verifyDefaultsXmlElementR_W_TypeHasValidValues() {
-        NodeList nodeList = XmlUtilities.getNodeList("//R_W_Type", Paths.get(path, "Metadata", "Defaults.xml"));
+        NodeList nodeList = XMLUtils.getNodeList("//R_W_Type", Paths.get(path, "Metadata", "Defaults.xml"));
 
         ArrayList<String> values = new ArrayList<>();
         List<String> VALID_VALUES = Arrays.asList("W", "R");
@@ -64,7 +63,7 @@ public class DefaultsXmlStructureTests extends CommonFixture {
 
     @Test
     public void verifyDefaultsXmlElementTypeHasValidValue() {
-        NodeList nodeList = XmlUtilities.getNodeList("//Type", Paths.get(path, "Metadata", "Defaults.xml"));
+        NodeList nodeList = XMLUtils.getNodeList("//Type", Paths.get(path, "Metadata", "Defaults.xml"));
 
         ArrayList<String> values = new ArrayList<>();
         List<String> VALID_VALUES = Arrays.asList("string", "integer", "float");
@@ -82,7 +81,7 @@ public class DefaultsXmlStructureTests extends CommonFixture {
     }
 
     private HashSet<String> getDatasetValues() {
-        NodeList datasetNodes = XmlUtilities.getNodeList("//Dataset", Paths.get(path, "Metadata", "Defaults.xml"));
+        NodeList datasetNodes = XMLUtils.getNodeList("//Dataset", Paths.get(path, "Metadata", "Defaults.xml"));
 
         HashSet<String> datasetValues = new HashSet<>();
 
@@ -94,7 +93,7 @@ public class DefaultsXmlStructureTests extends CommonFixture {
     }
 
     private ArrayList<String> collectNameNodesWithDatasetValue(String datasetValue) {
-        NodeList nameNodes = XmlUtilities.getNodeList("//Default_Value[Dataset/text() = \"" + datasetValue + "\"]/Name", Paths.get(path, "Metadata", "Defaults.xml"));
+        NodeList nameNodes = XMLUtils.getNodeList("//Default_Value[Dataset/text() = \"" + datasetValue + "\"]/Name", Paths.get(path, "Metadata", "Defaults.xml"));
 
         ArrayList<String> names = new ArrayList<>();
         for (int i = 0; i < nameNodes.getLength(); i++) {
