+++ src/main/java/org/opengis/cite/cdb10/MetadataAndVersioning/LightsXxxXmlStructureTests.java
@@ -2,7 +2,7 @@ package org.opengis.cite.cdb10.MetadataAndVersioning;
 
 import org.opengis.cite.cdb10.CommonFixture;
 import org.opengis.cite.cdb10.util.SchemaValidatorErrorHandler;
-import org.opengis.cite.cdb10.util.XmlUtilities;
+import org.opengis.cite.cdb10.util.XMLUtils;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import org.w3c.dom.Node;
@@ -60,7 +60,7 @@ public class LightsXxxXmlStructureTests extends CommonFixture {
         for (File xmlFile : getCustomLightsXmlFiles()) {
             File xsdFile = Paths.get(path, "Metadata", "Schema", "Lights_Tuning.xsd").toFile();
 
-            SchemaValidatorErrorHandler errorHandler = XmlUtilities.validateXmlFileIsValid(xmlFile, xsdFile);
+            SchemaValidatorErrorHandler errorHandler = XMLUtils.validateXmlFileIsValid(xmlFile, xsdFile);
 
             if (!errorHandler.noErrors()) {
                 Assert.fail(xmlFile.getName() + " does not contain valid XML. Errors: " + errorHandler.getMessages());
@@ -163,7 +163,7 @@ public class LightsXxxXmlStructureTests extends CommonFixture {
     }
 
     private ArrayList<String> getDirectionalityValues(File xmlFile) {
-        NodeList nodeList = XmlUtilities.getNodeList("//Directionality", Paths.get(path, "Metadata", xmlFile.getName()));
+        NodeList nodeList = XMLUtils.getNodeList("//Directionality", Paths.get(path, "Metadata", xmlFile.getName()));
 
         ArrayList<String> directionalityValues = new ArrayList<>();
 
@@ -175,7 +175,7 @@ public class LightsXxxXmlStructureTests extends CommonFixture {
     }
 
     private ArrayList<String> getInvalidColorValues(File xmlFile) {
-        NodeList frequencyNodes = XmlUtilities.getNodeList("//Color", Paths.get(path, "Metadata", xmlFile.getName()));
+        NodeList frequencyNodes = XMLUtils.getNodeList("//Color", Paths.get(path, "Metadata", xmlFile.getName()));
 
         ArrayList<String> invalidColorValues = new ArrayList<>();
 
@@ -194,7 +194,7 @@ public class LightsXxxXmlStructureTests extends CommonFixture {
     }
 
     private ArrayList<String> getInvalidPercentageValues(File xmlFile, String nodeToSearchFor) {
-        NodeList nodes = XmlUtilities.getNodeList(nodeToSearchFor, Paths.get(path, "Metadata", xmlFile.getName()));
+        NodeList nodes = XMLUtils.getNodeList(nodeToSearchFor, Paths.get(path, "Metadata", xmlFile.getName()));
 
         ArrayList<String> invalidValues = new ArrayList<>();
 
@@ -210,7 +210,7 @@ public class LightsXxxXmlStructureTests extends CommonFixture {
     }
 
     private ArrayList<String> getInvalidFrequencyValues(File xmlFile) {
-        NodeList frequencyNodes = XmlUtilities.getNodeList("//Frequency", Paths.get(path, "Metadata", xmlFile.getName()));
+        NodeList frequencyNodes = XMLUtils.getNodeList("//Frequency", Paths.get(path, "Metadata", xmlFile.getName()));
 
         ArrayList<String> invalidFrequencyValues = new ArrayList<>();
